adding Context.setCacheConfig. see here [1-2]. mac build.
[1] http://www.cs.cmu.edu/afs/cs/academic/class/15668-s11/www/cuda-doc/html/group__CUDA__EXEC_g40f8c11e81def95dc0072a375f965681.html#g40f8c11e81def95dc0072a375f965681
[2] http://developer.download.nvidia.com/compute/cuda/CUDA_Occupancy_calculator.xls

+++ src/org/trifort/rootbeer/generate/opencl/CudaHeader.c
@@ -34,7 +34,7 @@ int getBlockDimx(){
 }
 
 __device__
-int getGridDimx(){
+long long getGridDimx(){
   return gridDim.x;
 }
